Update callback arg
Updated the object passed to the user's callback when diffing is complete.
We now pass pack actual canvas objects of the source images instead of the
internal-only ImgWrapper class, which the end-user would never need.

Also updated the name of the parameter to better match what it actually did.

Updated README and tests to match new API.